refactor: sealed class Cell 을 상속받는 Empty, Mine 클래스를 동일한 파일내에 위치하도록 변경
