+++ src/main/kotlin/minesweeper/domain/board/NearbyMineCounter.kt
@@ -1,8 +1,8 @@
 package minesweeper.domain.board
 
+import minesweeper.domain.cell.Cell.Empty
+import minesweeper.domain.cell.Cell.Mine
 import minesweeper.domain.cell.Cells
-import minesweeper.domain.cell.Empty
-import minesweeper.domain.cell.Mine
 
 object NearbyMineCounter {
