--- src/main/kotlin/minesweeper/domain/cell/Mine.kt
@@ -1,3 +1,0 @@
-package minesweeper.domain.cell
-
-class Mine(position: Position) : Cell(position)
